refactor(frontend): restructure routing and layouts
- Move account-related pages into `accounts` folder
- Rename and reorganize account pages for clarity
- Add `AccountsLayout` and `GigsLayout` components for nested routing
- Update `AppRoutes` to use new layout-based route structure
- Add new `uploadGigsPage` under gigs feature
- Update forms and navbar to fit new structure
