+++ client/src/layouts/accountsLayout.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from "react-router-dom";
+
+const AccountsLayout = () => {
+  return (
+    <div>
+      <Outlet />
+    </div>
+  );
+};
+
+export default AccountsLayout;
