--- client/src/pages/ResetPasswordPage.tsx
@@ -1,12 +1,0 @@
-import ResetPasswordForm from "../components/forms/ResetPasswordForm";
-
-//TODO: add styling
-const ResetRequestPage = () => {
-  return (
-    <div>
-      <ResetPasswordForm /> 
-    </div>
-  )   
-}
-
-export default ResetRequestPage;
