+++ client/src/pages/accounts/ResetPasswordPage.tsx
@@ -1,0 +1,11 @@
+import ResetPasswordForm from "../../components/forms/ResetPasswordForm";
+
+const ResetPasswordPage = () => {
+  return (
+    <div>
+      <ResetPasswordForm /> 
+    </div>
+  )   
+}
+
+export default ResetPasswordPage;
