+++ client/src/pages/accounts/ResetRequestPage.tsx
@@ -1,6 +1,5 @@
-import ResetRequestForm from "../components/forms/ResetRequestForm";
+import ResetRequestForm from "../../components/forms/ResetRequestForm";
 
-//TODO: add styling
 const ResetRequestPage = () => {
   return (
     <div>
