+++ client/src/routes/AppRoutes.tsx
@@ -1,20 +1,29 @@
 import { Route, Routes } from "react-router-dom";
-import LoginPage from "../pages/LoginPage";
-import SignupPage from "../pages/SignupPage";
-import ResetRequestPage from "../pages/ResetRequestPage";
-import ResetPasswordPage from "../pages/ResetPasswordPage";
-import App from "../App"
+import App from "../App";
+import AccountsLayout from "../layouts/accountsLayout";
+import GigsLayout from "../layouts/gigsLayout";
+import LoginPage from "../pages/accounts/LoginPage";
+import ResetPasswordPage from "../pages/accounts/ResetPasswordPage";
+import ResetRequestPage from "../pages/accounts/ResetRequestPage";
+import SignupPage from "../pages/accounts/SignupPage";
 
 const AppRoutes = () => {
   return (
     <Routes>
-      <Route path="/" element={<App />} />
-      <Route path="/login" element={<LoginPage />} />
-      <Route path="/signup" element={<SignupPage />} />
-      <Route path="/request-reset" element={<ResetRequestPage />} />
-      <Route path="/reset-password" element={<ResetPasswordPage />} />
+      <Route index element={<App />} />
+
+      <Route path="accounts" element={<AccountsLayout />}>
+        <Route path="login" element={<LoginPage />} />
+        <Route path="signup" element={<SignupPage />} />
+        <Route path="request-reset" element={<ResetRequestPage />} />
+        <Route path="reset-password" element={<ResetPasswordPage />} />
+      </Route>
+
+      <Route path="gigs" element={<GigsLayout />}>
+        <Route path="upload" element={<div>upload gig form</div>} />
+      </Route>
     </Routes>
-  )
-}
+  );
+};
 
 export default AppRoutes;
